# Use `rerun::Collection` almost everywhere we'd use `std::vector` before