--- a/docs/code-examples/line_strip2d_batch.cpp
+++ b/docs/code-examples/line_strip2d_batch.cpp
@@ -13,7 +13,8 @@ int main() {
         {{0.f, 3.f}, {1.f, 4.f}, {2.f, 2.f}, {3.f, 4.f}, {4.f, 2.f}, {5.f, 4.f}, {6.f, 3.f}};
     rec.log(
         "strips",
-        rerun::LineStrips2D({strip1, strip2})
+        // TODO: Figure out how to avoid the extra vector
+        rerun::LineStrips2D(std::vector{strip1, strip2})
             .with_colors({0xFF0000FF, 0x00FF00FF})
             .with_radii({0.025f, 0.005f})
             .with_labels({"one strip here", "and one strip there"})