--- a/rerun_cpp/src/rerun/archetypes/asset3d.cpp
+++ b/rerun_cpp/src/rerun/archetypes/asset3d.cpp
@@ -19,28 +19,32 @@ namespace rerun {
         cells.reserve(3);
 
         {
-            auto result = Collection<rerun::components::Blob>(archetype.blob).serialize();
+            const size_t size = 1;
+            auto result = rerun::components::Blob::to_data_cell(&archetype.blob, size);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value));
+            cells.emplace_back(std::move(result.value), size);
         }
         if (archetype.media_type.has_value()) {
+            const size_t size = 1;
             auto result =
-                Collection<rerun::components::MediaType>(archetype.media_type.value()).serialize();
+                rerun::components::MediaType::to_data_cell(&archetype.media_type.value(), size);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value));
+            cells.emplace_back(std::move(result.value), size);
         }
         if (archetype.transform.has_value()) {
-            auto result =
-                Collection<rerun::components::OutOfTreeTransform3D>(archetype.transform.value())
-                    .serialize();
+            const size_t size = 1;
+            auto result = rerun::components::OutOfTreeTransform3D::to_data_cell(
+                &archetype.transform.value(),
+                size
+            );
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value));
+            cells.emplace_back(std::move(result.value), size);
         }
         {
-            auto result =
-                Collection<Asset3D::IndicatorComponent>(Asset3D::IndicatorComponent()).serialize();
+            auto indicator = Asset3D::IndicatorComponent();
+            auto result = Asset3D::IndicatorComponent::to_data_cell(&indicator, 1);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value));
+            cells.emplace_back(std::move(result.value), 1);
         }
 
         return cells;