--- a/rerun_cpp/src/rerun/archetypes/asset3d.hpp
+++ b/rerun_cpp/src/rerun/archetypes/asset3d.hpp
@@ -4,6 +4,7 @@
 #pragma once
 
 #include "../collection.hpp"
+#include "../compiler_utils.hpp"
 #include "../components/blob.hpp"
 #include "../components/media_type.hpp"
 #include "../components/out_of_tree_transform3d.hpp"
@@ -11,7 +12,6 @@
 #include "../indicator_component.hpp"
 #include "../result.hpp"
 #include "../serialized_component_batch.hpp"
-#include "../warning_macros.hpp"
 
 #include <cstdint>
 #include <filesystem>
@@ -118,7 +118,7 @@ namespace rerun::archetypes {
         Asset3D with_media_type(rerun::components::MediaType _media_type) && {
             media_type = std::move(_media_type);
             // See: https://github.com/rerun-io/rerun/issues/4027
-            WITH_MAYBE_UNINITIALIZED_DISABLED(return std::move(*this);)
+            RERUN_WITH_MAYBE_UNINITIALIZED_DISABLED(return std::move(*this);)
         }
 
         /// An out-of-tree transform.
@@ -127,7 +127,7 @@ namespace rerun::archetypes {
         Asset3D with_transform(rerun::components::OutOfTreeTransform3D _transform) && {
             transform = std::move(_transform);
             // See: https://github.com/rerun-io/rerun/issues/4027
-            WITH_MAYBE_UNINITIALIZED_DISABLED(return std::move(*this);)
+            RERUN_WITH_MAYBE_UNINITIALIZED_DISABLED(return std::move(*this);)
         }
 
         /// Returns the number of primary instances of this archetype.