--- a/rerun_cpp/src/rerun/archetypes/asset3d_ext.cpp
+++ b/rerun_cpp/src/rerun/archetypes/asset3d_ext.cpp
@@ -12,8 +12,7 @@
 
 // </CODEGEN_COPY_TO_HEADER>
 
-namespace rerun {
-    namespace archetypes {
+namespace rerun::archetypes {
 
 #if 0
         // <CODEGEN_COPY_TO_HEADER>
@@ -35,10 +34,10 @@ namespace rerun {
         /// If no `MediaType` is specified, the Rerun Viewer will try to guess one from the data
         /// at render-time. If it can't, rendering will fail with an error.
         static Asset3D from_bytes(
-            const std::vector<uint8_t> bytes, std::optional<rerun::components::MediaType> media_type
+            rerun::Collection<uint8_t> bytes, std::optional<rerun::components::MediaType> media_type
         ) {
             // TODO(cmc): we could try and guess using magic bytes here, like rust does.
-            Asset3D asset = Asset3D(bytes);
+            Asset3D asset = Asset3D(std::move(bytes));
             asset.media_type = media_type;
             return asset;
         }
@@ -46,39 +45,40 @@ namespace rerun {
         // </CODEGEN_COPY_TO_HEADER>
 #endif
 
-        Result<Asset3D> Asset3D::from_file(const std::filesystem::path& path) {
-            std::ifstream file(path, std::ios::binary);
-            if (!file) {
-                return Error(ErrorCode::FileOpenFailure, "Failed to open file: " + path.string());
-            }
-
-            file.seekg(0, std::ios::end);
-            std::streampos length = file.tellg();
-            file.seekg(0, std::ios::beg);
-
-            std::vector<uint8_t> data(static_cast<size_t>(length));
-            file.read(reinterpret_cast<char*>(data.data()), length);
-
-            return Asset3D::from_bytes(data, Asset3D::guess_media_type(path));
+    Result<Asset3D> Asset3D::from_file(const std::filesystem::path& path) {
+        std::ifstream file(path, std::ios::binary);
+        if (!file) {
+            return Error(ErrorCode::FileOpenFailure, "Failed to open file: " + path.string());
         }
 
-        std::optional<rerun::components::MediaType> Asset3D::guess_media_type(
-            const std::filesystem::path& path
-        ) {
-            std::filesystem::path file_path(path);
-            std::string ext = file_path.extension().string();
-            std::transform(ext.begin(), ext.end(), ext.begin(), ::tolower);
+        file.seekg(0, std::ios::end);
+        std::streampos length = file.tellg();
+        file.seekg(0, std::ios::beg);
 
-            if (ext == ".glb") {
-                return rerun::components::MediaType::glb();
-            } else if (ext == ".gltf") {
-                return rerun::components::MediaType::gltf();
-            } else if (ext == ".obj") {
-                return rerun::components::MediaType::obj();
-            } else {
-                return std::nullopt;
-            }
+        std::vector<uint8_t> data(static_cast<size_t>(length));
+        file.read(reinterpret_cast<char*>(data.data()), length);
+
+        return Asset3D::from_bytes(
+            Collection<uint8_t>::borrow(data.data(), data.size()),
+            Asset3D::guess_media_type(path)
+        );
+    }
+
+    std::optional<rerun::components::MediaType> Asset3D::guess_media_type(
+        const std::filesystem::path& path
+    ) {
+        std::filesystem::path file_path(path);
+        std::string ext = file_path.extension().string();
+        std::transform(ext.begin(), ext.end(), ext.begin(), ::tolower);
+
+        if (ext == ".glb") {
+            return rerun::components::MediaType::glb();
+        } else if (ext == ".gltf") {
+            return rerun::components::MediaType::gltf();
+        } else if (ext == ".obj") {
+            return rerun::components::MediaType::obj();
+        } else {
+            return std::nullopt;
         }
-
-    } // namespace archetypes
-} // namespace rerun
+    }
+} // namespace rerun::archetypes