--- a/rerun_cpp/src/rerun/archetypes/bar_chart.cpp
+++ b/rerun_cpp/src/rerun/archetypes/bar_chart.cpp
@@ -19,15 +19,16 @@ namespace rerun {
         cells.reserve(1);
 
         {
-            auto result = Collection<rerun::components::TensorData>(archetype.values).serialize();
+            const size_t size = 1;
+            auto result = rerun::components::TensorData::to_data_cell(&archetype.values, size);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value));
+            cells.emplace_back(std::move(result.value), size);
         }
         {
-            auto result = Collection<BarChart::IndicatorComponent>(BarChart::IndicatorComponent())
-                              .serialize();
+            auto indicator = BarChart::IndicatorComponent();
+            auto result = BarChart::IndicatorComponent::to_data_cell(&indicator, 1);
             RR_RETURN_NOT_OK(result.error);
-            cells.emplace_back(std::move(result.value));
+            cells.emplace_back(std::move(result.value), 1);
         }
 
         return cells;